Program statement case insensitivity and ending keyword
Pascal is a case insensitve language. We can use a mode modifier
`(?i)` to switch off case sensitivity for the rest of the
expression. `(?i)(program)` will match `program`, `PROGRAM`,
`PrOgRaM`, and any other combination of character casing.

Additionally the `end.` keyword is included as the end of a
program statement.