Remove negated hyphen from identifier regex
The part `[^-]*` was too difficult to express and was causing
non-matches. A Pascal identifier may start with an underscore,
or a letter. The rest may be any digit or character except
hyphen. For the sake of simplicity I've removed matching a
negated hyphen.

Reference: https://www.freepascal.org/docs-html/current/ref/refse4.html#x15-140001.4