Add if statements to programs and units
It is far easier to treat if-then and else as two separate
statements. It gets very hard to create a regex that matches all
the possible statements that both branches may contain.

Also it is very difficult to match the range of expressions that
are possible in the boolean test so I am just matching all
characters. This is naive but far quicker and easier to write a
regex for.